move and rename new functions; response to review comments
