+++ py/torch_migraphx/fx/converters/aten_ops_converters.py
@@ -297,7 +297,7 @@ def aten_ops_clamp(mgx_module, node, args, _kwargs):
 
 
 @migraphx_converter(torch.ops.aten.nll_loss_forward.default)
-def aten_ops_nllloss(mgx_module, node, args, _kwargs):
+def aten_ops_nll_loss(mgx_module, node, args, _kwargs):
     assert len(args) == 5
 
     #TODO: find where this enum comversion is defined in library code
@@ -329,7 +329,7 @@ def aten_ops_nllloss(mgx_module, node, args, _kwargs):
     weights_inst  = mgx_module.add_literal(torch.tensor((targets), dtype=torch.int64).numpy())
     
     # The acc op returns one tensor, while the aten op returns two tensors
-    return acc_ops_converters.acc_ops_nll_loss_forward(mgx_module, node, (), acc_kwargs), MGXInstruction(weights_inst)
+    return acc_ops_converters.acc_ops_nll_loss(mgx_module, node, (), acc_kwargs), MGXInstruction(weights_inst)
 
 
 @migraphx_converter(torch.ops.aten.relu.default)
